finished filter visitors implementation

+++ commercetools.Sdk/commercetools.Sdk/Product/Attributes/NumberAttribute.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 namespace commercetools.Sdk.Domain
 {
-    public class NumberAttribute : Attribute<int>
+    public class NumberAttribute : Attribute<double>
     {
         // TODO See which type should be used exactly
     }
